refactor: make listOrders queryparams extracting to its own helper function.
