+++ server/src/utils/request-helpers.ts
@@ -1,4 +1,5 @@
 // utils/requests-helpers.ts
+import { Request } from "express";
 
 // Extracts filters from query params
 export const extractFilters = (
@@ -11,6 +12,29 @@ export const extractFilters = (
   }, {} as Record<string, string>);
 };
 
+// Extracts and normalizes order query parameters from a request.
+export const extractOrderQueryParams = (req: Request) => {
+  const { search, sort, page, limit, startDate, endDate } = req.query;
+
+  const filters = extractFilters(req.query);
+  delete filters.startDate;
+  delete filters.endDate;
+
+  const { page: parsedPage, limit: parsedLimit } = parsePagination(
+    page?.toString(),
+    limit?.toString()
+  );
+
+  return {
+    search: search ? search.toString() : "",
+    sort: sort ? sort.toString() : "",
+    page: parsedPage,
+    limit: parsedLimit,
+    filters,
+    startDate: startDate ? startDate.toString() : undefined,
+    endDate: endDate ? endDate.toString() : undefined,
+  };
+};
 
 // Parses query params for pagination.
 export const parsePagination = (
